% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_treatment_id.R
\name{propagate_treatment_id}
\alias{propagate_treatment_id}
\title{Assign id to leads and lags.}
\usage{
propagate_treatment_id(time, id, treatment_id, window)
}
\arguments{
\item{time}{A column name containing the time variable in the panel (i.e. the t dimension)}

\item{id}{A column name containing the unique identifier of the observational unit in the panel (the i dimension)}

\item{treatment_id}{A character specifying the treatment id, as produced by \code{treatment_ids()}}

\item{window}{A number indicating the number of periods within which two or more treatment events are merged. This should be in the same units as \code{time}}
}
\value{
A character vector specifiying the treatment id for treated and control periods of a given unit
}
\description{
Propagates the treatment id from \code{treatment_ids} across control periods
}
\examples{
# Load packages
library(tidyr)
library(dplyr)
library(EventHorizon)

# Load a panel
panel <- simulate_panel(n_units = 10, n_periods = 10)
head(panel)

event_horizon_panel <- panel |>
mutate(treatment_id = treatment_ids(id = id,
                                    time = time,
                                    treatment = treatment,
                                    window = 3),
       relative_time = calculate_relative_time(id = id,
                                               time = time,
                                               treatment_id = treatment_id,
                                               window = 3),
       treatment_id = propagate_treatment_id(id = id,
                                             time = time,
                                             treatment_id = treatment_id,
                                             window = 3)) |>
drop_na(treatment_id)

head(event_horizon_panel)
visualize_panel(panel)
visualize_panel(event_horizon_panel)
visualize_panel(event_horizon_panel, prepped = TRUE)
}
