% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_ids.R
\name{treatment_ids}
\alias{treatment_ids}
\title{Identify treatment ids based on a user-defined window}
\usage{
treatment_ids(id, time, treatment, window)
}
\arguments{
\item{id}{A column name containing the unique identifier of the observational unit in the panel (the i dimension)}

\item{time}{A column name containing the time variable in the panel (i.e. the t dimension)}

\item{treatment}{A column name containing the name of the dummy variable indicating whether unit i is treated at time t}

\item{window}{A number indicating the number of periods within which two or more treatment events are merged. This should be in the same units as \code{time}}
}
\value{
A character vector specifying the new treatment ids
}
\description{
If in treatment falls within the window, it is considered a single treatment event and a new treatment id is created.
}
\examples{
# Load a panel
panel <- simulate_panel(n_units = 10, n_periods = 10)
head(panel)

event_horizon_panel <- panel |>
mutate(treatment_id = treatment_ids(id = id, time = time, treatment = treatment, window = 3))

head(event_horizon_panel)

}
