% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_panel.R
\name{visualize_panel}
\alias{visualize_panel}
\title{Visualize a panel}
\usage{
visualize_panel(data, id = id, time = time, treatment = treatment, prepped = F)
}
\arguments{
\item{data}{An object of class data.frame or tibble containing the panel data}

\item{id}{A column name containing the unique identifier of the observational unit in the panel (the i dimension)}

\item{time}{A column name containing the time variable in the panel (i.e. the t dimension)}

\item{treatment}{A column name containing the name of the dummy variable indicating whether unit i is treated at time t}

\item{prepped}{A boolean specifying if the panel has already been prepared through this package (i.e. the treatment_ids have been created and modified). If so, the points will be colored based on treatment id.}
}
\value{
a ggplot object
}
\description{
Visualize a panel
}
\examples{
# Load packages
library(tidyr)
library(dplyr)
library(EventHorizon)

panel <- simulate_panel(n_units = 10, n_periods = 10)
visualize_panel(panel)

event_horizon_panel <- panel |>
mutate(treatment_id = treatment_ids(id = id,
                                    time = time,
                                    treatment = treatment,
                                    window = 3),
       relative_time = calculate_relative_time(id = id,
                                               treatment_id = treatment_id,
                                               window = 3),
       treatment_id = propagate_treatment_id(id = id,
                                             time = time,
                                             treatment_id = treatment_id,
                                             window = 3)) |>
drop_na(treatment_id)

visualize_panel(event_horizon_panel)
visualize_panel(event_horizon_panel, prepped = TRUE)
}
